ethica-framework/requirements.txt
google-generativeai>=0.3.0
mistralai>=0.1.0
requests>=2.31.0
python-dotenv>=1.0.0
